+++ database/seeders/database_seeder.go
@@ -1,9 +1,0 @@
-package seeders
-
-import "github.com/zatrano/framework/packages/database/seeder"
-
-// All returns application seeders.
-// Auth/dashboard seed data is added by make:auth / make:dashboard (or your own seeders).
-func All() []seeder.Seeder {
-	return []seeder.Seeder{}
-}